refactor: inputbar view와 로직 파일 분리
