+++ components/main/transactionsList.js
@@ -4,7 +4,10 @@ import {
 } from "../../utils/transaction.js";
 import { CATEGORY_NAME } from "../../constants/category.js";
 import { formatMoney } from "../../utils/format.js";
-import { fillFormWithTransaction, cancelEditMode } from "./inputBar.js";
+import {
+  fillFormWithTransaction,
+  cancelEditMode,
+} from "./inputBar.viewmodel.js";
 import { dateStore, transactionStore } from "../../store/index.js";
 
 // 클릭된 행 상태 관리
