+++ constants/form.js
@@ -1,0 +1,7 @@
+export const FORM_FIELDS = {
+  DATE: "date",
+  AMOUNT: "amount",
+  DESCRIPTION: "description",
+  PAYMENT_METHOD: "paymentMethod",
+  CATEGORY: "category",
+};
